+++ CHANGELOG.md
@@ -1,6 +1,37 @@
 # Change Log 
 ## [Unreleased]
 
+## Released: [1.0] - 2022-03-23
+## [1.0] - 2022-03-23
+## Added
+- [`README.md`] lines 13-24: added `## Scripts` and `## Compatibility`
+
+## [1.0] - 2022-03-23
+## Added
+- [`index.html`] lines 69-71: added title image link
+
+## [1.0] - 2022-03-20
+## Added
+- [`README.md`] lines 9-10: added title image link
+
+## [1.0] - 2022-03-18
+## Added
+- [`index.html`] line 53: `<br><code>(0-9,A-F,a-f)</code>`
+
+## Fixed
+- [`index_main.js`] line 52:  
+  ```
+   		arcLandscape(cId,INPUT.value.replaceAll(/[^0-1A-Fa-f]/g,''),!REVERSE.checked);
+  ```
+
+## Added
+- [`index_main.js`] lines 74 and 75: added a comment to show output filename (`"AL<number>_TYPE<type|typeReverse><width>x<height>.png"`)
+
+## [1.0] - 2022-03-17
+## Changed
+- [`index_main.js`]: lines 21-23, 29-31, 37-39, 45-47, 53-55 and 74: changed filename of output to download
+- [`index.html`] line 31: a link to wiki page
+
 ## Released: [0.1 beta] - 2022-03-17
 ## [0.1 beta] - 2022-03-16
 ## Fixed